Refactor sessionlog related things
Split repository
Implement update and delete functionality in controller
Delete create view
