+++ Console/CodingTracker/CodingTracker/Views/SessionLogView.cs
@@ -4,13 +4,13 @@ namespace CodingTracker.Views;
 
 public class SessionLogView : BaseView
 {
-    public SessionLog AskSessionTimes()
+    public SessionLog AskSessionTimes(SessionLog? toUpdateLog = null)
     {
         var errorMessage = "[red]Invalid time format. Use 24-hour format (HH:mm).[/]";
-        var startTime = AskInput<string>($"At what time did you start? [grey]({Validator.TimeFormat})[/]",
-            Validator.ValidateStringAsTime, errorMessage);
-        var endTime = AskInput<string>($"At what time did you end? [grey]({Validator.TimeFormat})[/]",
-            Validator.ValidateStringAsTime, errorMessage);
+        var startTime = AskInput($"At what time did you start? [grey]({Validator.TimeFormat})[/]",
+            Validator.ValidateStringAsTime, errorMessage, toUpdateLog?.StartTime.ToString(Validator.TimeFormat));
+        var endTime = AskInput($"At what time did you end? [grey]({Validator.TimeFormat})[/]",
+            Validator.ValidateStringAsTime, errorMessage, toUpdateLog?.EndTime.ToString(Validator.TimeFormat));
 
         return new SessionLog
         {
